+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 1.12.0 (TBD)
+
+### Behavior change
+
+- Added two optional arguments to `Session.add_import`.
+  - `chunk_size`: The number of bytes to hash per chunk of the uploaded files.
+  - `whole_file_hash`: By default only the first chunk of the uploaded import is hashed to save time. When this is set to True each uploaded file is fully hashed instead.
+
 ## 1.11.1 (2023-11-07)
 
 ### Bug Fixes